Fix block label not rendered by default
